Ispravljene su metode za prikazivanje i brisanje requestova

+++ src/main/java/com/example/SocialNetwork/controller/MembershipRequestController.java
@@ -1,8 +1,10 @@
 package com.example.SocialNetwork.controller;
 
 import com.example.SocialNetwork.entities.MembershipRequest;
+import com.example.SocialNetwork.entities.RequestStatus;
 import com.example.SocialNetwork.entities.SocialGroup;
 import com.example.SocialNetwork.entities.User;
+import com.example.SocialNetwork.helpercalsses.MyRequest;
 import com.example.SocialNetwork.service.MembershipRequestService;
 import com.example.SocialNetwork.service.SocialGroupService;
 import com.example.SocialNetwork.service.UserService;
@@ -16,8 +18,6 @@ public class MembershipRequestController {
     private final MembershipRequestService requestService;
     private final UserService userService;
     private final SocialGroupService groupService;
-    private Long id;
-
     public MembershipRequestController(MembershipRequestService requestService, UserService userService, SocialGroupService groupService){
         this.requestService = requestService;
         this.userService = userService;
@@ -29,25 +29,49 @@ public class MembershipRequestController {
         User user = userService.findCurrentUser();
         SocialGroup socialGroup = groupService.getSocialGroupById(id);
         if(socialGroup.getUser().getId()==user.getId()){
-            return requestService.getAllRequests();
+            requestService.getAllRequestsForSocialGroup(id);
         }
         return  null;
     }
 
     @GetMapping("/{id}")
     public MembershipRequest getRequestsById(@PathVariable Long id) {
-        return requestService.getAllRequestsById(id);
+        User currentUser = userService.findCurrentUser();
+        MembershipRequest request = requestService.getAllRequestsById(id);
+        if (request.getUser().getId() == currentUser.getId() || request.getSocialGroup().getUser().getId()==currentUser.getId())
+            return request;
+        else
+            return null;
     }
 
+    @PostMapping("/")
+    public String createMembershipRequest(@RequestBody MyRequest id) {
+        SocialGroup socialGroup = groupService.getSocialGroupById(id.getId());
+        User u = userService.findCurrentUser();
+        if(u!=null){
+            MembershipRequest membershipRequest = new MembershipRequest();
+            membershipRequest.setSocialGroup(socialGroup);
+            membershipRequest.setUser(u);
+            membershipRequest.setRequestStatus(RequestStatus.PENDING);
+            requestService.saveRequest(membershipRequest);
+            return "Uspesno ste poslali request za uclanjenje u grupu";
+        }
+        else
+            return "Zahteve mogu poslati samo ulogovani koristnici";
+
+    }
     @DeleteMapping("/{id}")
     public String deleteRequestById(@PathVariable Long id) {
-        requestService.deleteRequestById(id);
+        User currentUser = userService.findCurrentUser();
+        MembershipRequest request = requestService.getAllRequestsById(id);
+        if (request.getUser().getId() == currentUser.getId() || request.getSocialGroup().getUser().getId()==currentUser.getId()){
 
-        return "Usesno ste obrisali zahtev";
+            requestService.deleteRequestById(id);
+            return "Uspesno ste obrisali zahtev";
+        }
+        else
+            return "Nemate pravo na brisanje zahteva";
     }
 
-    @GetMapping("/allrequestsforgroup/{id}")
-    public List<MembershipRequest> showAllRequestsForSocialGroup(@PathVariable Long id){
-        return requestService.getAllRequestsForSocialGroup(id);
-    }
+
 }
